Move QuickEditItems to a subfolder
