+++ src/HexManiac.Core/HexManiac.Core.csproj
@@ -60,7 +60,7 @@
     <Compile Include="Models\Runs\PLMRun.cs" />
     <Compile Include="Models\Runs\WordRun.cs" />
     <Compile Include="ViewModels\AutoCompleteSelectionItem.cs" />
-    <Compile Include="ViewModels\IQuickEditItem.cs" />
+    <Compile Include="ViewModels\QuickEditItems\IQuickEditItem.cs" />
     <Compile Include="ViewModels\Tools\CodeTool.cs" />
     <Compile Include="ViewModels\Visitors\CompleteEditOperation.cs" />
     <Compile Include="ViewModels\Visitors\ContextItemFactory.cs" />
