+++ src/HexManiac.Tests/AutoSearchTests.cs
@@ -3,6 +3,7 @@ using HavenSoft.HexManiac.Core.Models;
 using HavenSoft.HexManiac.Core.Models.Runs;
 using HavenSoft.HexManiac.Core.ViewModels;
 using HavenSoft.HexManiac.Core.ViewModels.DataFormats;
+using HavenSoft.HexManiac.Core.ViewModels.QuickEditItems;
 using System;
 using System.Collections.Generic;
 using System.IO;
